Create SVM Model_ronak
Adding my SVM code here. I used under sampling instead of SMOTE and got better results. Think we should follow Lucy's suggestion and try to all use the same code for pre-processing so y'all could just copy the sections of code that are unique to ur model and put it in here to make it easier